Extracted Attachment from DataAttachmentsWidget
